pyproject.toml
[tool.poetry]
name = "cdd-agent"
version = "0.0.4"
description = "AI coding assistant with structured CDD workflows - LLM agnostic terminal agent"
authors = ["Guilherme Gouw <guilherme.gouw@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/guilhermegouw/context-driven-development-agent"
repository = "https://github.com/guilhermegouw/context-driven-development-agent"
keywords = ["ai", "llm", "coding-assistant", "anthropic", "openai", "cli", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
packages = [{include = "cdd_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
rich = "^13.7.0"
textual = "^0.87.1"
anthropic = "^0.40.0"
openai = "^1.55.0"
pydantic = "^2.10.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
authlib = "^1.3.2"
catppuccin = {extras = ["pygments"], version = "^2.5.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
mypy = "^1.13.0"
ruff = "^0.8.0"
pytest-cov = "^6.0.0"

[tool.poetry.scripts]
cdd-agent = "cdd_agent.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=cdd_agent --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
